Move main hecate server logic to lib
